% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aircraft_types}
\alias{aircraft_types}
\title{Aircraft type info}
\format{
A data frame with 322 rows and 15 columns:
\itemize{
\item \code{icao}: the ICAO aircraft type code.
\item \code{name}: the aircraft type name.
\item \code{manufacturer}: the manufacturer.
\item \code{type}: the type, i.e. L2J for \strong{L}andplane \strong{2} \strong{J}et engines craft
\item \code{wtc}: Wake Turbulence Category
\item \code{...}: the rest
}
}
\usage{
aircraft_types
}
\description{
Aircraft type info
}
\keyword{datasets}
