% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vallado.R
\name{coe2rv}
\alias{coe2rv}
\title{Find the position and velocity vectors in geocentric equatorial (ijk) system
given the classical orbit elements}
\usage{
coe2rv(p, ecc, incl, raan, argp, nu, arglat = NA, truelon = NA, lonper = NA)
}
\arguments{
\item{p}{semilatus rectum, `p`, [km]}

\item{ecc}{eccentricity, `e`, [unitless]}

\item{incl}{inclination, `i`, \eqn{[0, \pi]} [rad]}

\item{raan}{right ascension of ascending node, i.e. the longitude of
the ascending node, `\eqn{\Omega}`}

\item{argp}{argument of perigee, `\eqn{\omega}`, \eqn{[0, 2\pi]} [rad]}

\item{nu}{true anomaly, `\eqn{\nu}`, \eqn{[0, 2\pi]} [rad]}

\item{arglat}{argument of latitude (for circular inclined orbit),
\eqn{[0, 2\pi]} [rad]}

\item{truelon}{true longitude (for circular equatorial orbit),
\eqn{[0, 2\pi]} [rad]}

\item{lonper}{longitude of periapsis (for elliptical equatorial orbit),
\eqn{[0, 2\pi]} [rad]}
}
\value{
a list of 2 vectors, `r` [km] and `v` [km/s] with named components
         `i`, `j`, and `k`, and `vi`, `vj` and `vk` respectively.
}
\description{
The additional orbital elements provide calculations for perfectly circular
and equatorial orbits.
}
\examples{
\dontrun{
p <- 11067.790 |> units::set_units("km") # km
ecc <- 0.83285 |> units::as_units()
incl <- 87.87 |> units::set_units("degree")
raan <- 227.89 |> units::set_units("degree")
argp <- 53.38 |> units::set_units("degree")
nu <- 92.335 |> units::set_units("degree")

coe2rv(p, ecc, incl, raan, argp, nu)
}
}
