% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_and_casuality.R
\name{collision_and_casuality_risk_expectation_hourly}
\alias{collision_and_casuality_risk_expectation_hourly}
\title{Calculate hourly collision and casuality risk expectation}
\usage{
collision_and_casuality_risk_expectation_hourly(days, resolution = 3L)
}
\arguments{
\item{days}{dates to calculate for}

\item{resolution}{H3 hex resolution}
}
\value{
save the result in a parquet file `collision_and_casuality_hourly.parquet`
}
\description{
Calculate hourly collision and casuality risk expectation
}
