% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_and_casualty.R
\name{collision_and_casualty_risk_expectation_hourly}
\alias{collision_and_casualty_risk_expectation_hourly}
\title{Calculate hourly collision and casualty risk expectation}
\usage{
collision_and_casualty_risk_expectation_hourly(day, resolution = 3L)
}
\arguments{
\item{day}{date to calculate for}

\item{resolution}{H3 hex resolution}
}
\value{
a dataframe with hourly \code{collision_expectation} and
\code{casualty_risk}, saved in a parquet file
\code{collision_and_casualty_hourly.parquet}
}
\description{
Calculate hourly collision and casualty risk expectation
}
