% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vallado.R
\name{ecef2ll}
\alias{ecef2ll}
\title{convert a geocentric equatorial position vector into latitude and longitude}
\usage{
ecef2ll(r)
}
\arguments{
\item{r}{ECEF (I, J, K) position vector [km]}
}
\value{
a list of the following named values
         * `latgc`: geocentric latitude of satellite, \eqn{[-\pi/2, \pi/2]} [rad]
         * `latgd`: geodetic latitude of satellite, \eqn{[-\pi/2, \pi/2]} [rad]
         * `lon`: longitude of satellite, \eqn{[-\pi, \pi]} [rad]
         * `hellp`: height above the ellipsoid, [km]
}
\description{
This function finds both geodetic and geocentric latitude.
}
\examples{
\dontrun{
r <- c(
  i = 6524.834,
  j = 6862.875,
  k = 6448.296
) |>
  units::set_units("km")

res <- ecef2ll(r)
}
}
