% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effective_expose_area}
\alias{effective_expose_area}
\title{Effective Exposed Area}
\format{
## `effective_expose_area`
A data frame with columns:
* `icao`: the ICAO aircraft type code.
* `eea`: the effective exposed area.
* `pax`: the max number of passengers.
}
\usage{
effective_expose_area
}
\description{
Effective Exposed Area
}
\keyword{datasets}
