% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonize.R
\name{hexagonize_traffic}
\alias{hexagonize_traffic}
\title{assign H3 cell index at specified resolution to all trajectory point}
\usage{
hexagonize_traffic(
  day,
  resolution = 3L,
  interval = 30L,
  bbox = c(xmin = -40.01297, ymin = 16.99059, xmax = 46.76206, ymax = 82.00901)
)
}
\arguments{
\item{day}{the date for the trajectories, it refers to the relevant
parquet file in `data/trajectories_<YYYY-MM-DD>_resampled_<interval>s.parquet`}

\item{resolution}{the H3 resolution}

\item{interval}{resampling interval}

\item{bbox}{a bounding box with names `xmin`, `xmax`, `ymin` and `ymax`}
}
\value{
a new parquet file as
  `data/trajectories_<YYYY-MM-DD>_resampled_<interval>s_bbox_res_<resolution>.parquet`
}
\description{
Keep only portions in the BBOX of the study
}
