% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_spatial.R
\name{hexes_for_bbox_at_res}
\alias{hexes_for_bbox_at_res}
\title{H3 hexagons covering a bounding box.}
\usage{
hexes_for_bbox_at_res(bbox, resolution)
}
\arguments{
\item{bbox}{bounding box to be covered by H3 hexagons}

\item{resolution}{H3 hexagons' resolution}
}
\value{
simple features of H3 hexagons polygons and relevant addresses
}
\description{
H3 hexagons covering a bounding box.
}
