% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbitals.R
\name{kepler_to_cartesian}
\alias{kepler_to_cartesian}
\title{convert from orbital elements to Cartesian coordinates}
\usage{
kepler_to_cartesian(nu, a, e, i, Omega, w, M = 5.9722e+24, m = 0)
}
\arguments{
\item{nu}{the true anomaly, i.e. the angular displacement measured from
periapsis to the position vector along the direction of motion
[radians]}

\item{a}{the semi-major axis of the orbit [m]}

\item{e}{the eccentricity of the orbit}

\item{i}{the inclination of the orbit [radians]}

\item{Omega}{the Right ascension of the ascending node [radians]}

\item{w}{the argument of perigee \[radians\]}

\item{M}{the mass of the first body (more massive one), i.e. Earth, [kg]}

\item{m}{the mass of the orbiting body (less massive), i.e. satellite, [kg].
In the case of m << M, m can be passed as 0}
}
\value{
a vector with I, J, K, Vi, Vj, Vk of the values of position and
         velocity in a geocentric-equatorial reference system
}
\description{
convert from orbital elements to Cartesian coordinates
}
\examples{
\dontrun{
rE  <- 6378e3 # Earth equatorial radius
alt <- 550e3 # altitude of satellite, i.e. 500 km for LEO
ecc <- 0 # circular orbit, i.e. when uncontrolled reentry of debris
m   <- 0 # mass of debris/satellite << mass Earth, so it can be 0
o   <- 0
w   <- 0
i <- 10
kepler_to_cartesian(
  nu = 0.5,
  a = rE + alt,
  e = ecc,
  i = i, Omega = o,
  w = w,
  m = m)
}
}
