% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbitals.R
\name{latitude_weights}
\alias{latitude_weights}
\title{latitude weight for a space object circularly orbiting at `inclination`}
\usage{
latitude_weights(delta_lat, altitude, inclination_deg, n = 10001L)
}
\arguments{
\item{delta_lat}{delta altitude [decimal degrees]}

\item{altitude}{altitude of the orbiting space object [m]}

\item{inclination_deg}{inclination [decimal degrees]}

\item{n}{number of samples of the orbit}
}
\value{
a tibble of latitudes and values for the density
}
\description{
latitude weight for a space object circularly orbiting at `inclination`
}
\examples{
\dontrun{
rE  <- 6378e3 # Earth equatorial radius
alt <- 550e3 # altitude of satellite, i.e. 500 km for LEO
latitude_weights(0.5, rE + alt, 50)
}
}
