% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_spatial.R
\name{plot_hexes_map}
\alias{plot_hexes_map}
\title{Plot H3 hex map}
\usage{
plot_hexes_map(resolution)
}
\arguments{
\item{resolution}{cells resolution}
}
\value{
a plot
}
\description{
Plot H3 hex map
}
\examples{
\dontrun{
plot_hexes_map(2L)
}
}
