% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vallado.R
\name{rot1}
\alias{rot1}
\title{Rotate a vector about the 1st axis}
\usage{
rot1(vec, theta)
}
\arguments{
\item{vec}{a 3D vector to rotate}

\item{theta}{the angle of rotation [rad]}
}
\value{
a 3D vector
}
\description{
Rotate a vector about the 1st axis
}
\examples{
\dontrun{
c(0, 1, 0) |> rot1(pi/2)
}
}
