% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vallado.R
\name{rot3}
\alias{rot3}
\title{Rotate a vector about the 3rd axis}
\usage{
rot3(vec, theta)
}
\arguments{
\item{vec}{a 3D vector to rotate}

\item{theta}{the angle of rotation [rad]}
}
\value{
a 3D vector
}
\description{
Rotate a vector about the 3rd axis
}
\examples{
\dontrun{
c(0, 1, 0) |> rot3(-pi/2)
}
}
