% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aircraft_performance_scrapers.R
\name{scrape_aircraft_type_info_skybrary}
\alias{scrape_aircraft_type_info_skybrary}
\title{Scraper for EUROCONTROL's Skybrary aircraft info}
\usage{
scrape_aircraft_type_info_skybrary(ac_type, session)
}
\arguments{
\item{ac_type}{ICAO aircraft type}

\item{session}{\code{polite} session}
}
\value{
a dataframe of aircraft info
}
\description{
Scraper for EUROCONTROL's Skybrary aircraft info
}
\examples{
\dontrun{
host <- "https://skybrary.aero/"
session <- polite::bow(host, force = TRUE)
scraper <- purrr::partial(scrape_aircraft_type_info_skybrary, session = session)

 ac_types |>
  map(.f = scraper) |>
  bind_rows()
}
}
