% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aircraft_types}
\alias{aircraft_types}
\title{Aircraft type info}
\format{
## `aircraft_types`
A data frame with 322 rows and 15 columns:
* `icao`: the ICAO aircraft type code.
* `name`: the aircraft type name.
* `manufacturer`: the manufacturer.
* `type`: the type, i.e. L2J for **L**andplane **2** **J**et engines craft
* `wtc`: Wake Turbulence Category
* `...`: the rest
}
\usage{
aircraft_types
}
\description{
Aircraft type info
}
\keyword{datasets}
