% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aviodebris-package.R
\docType{package}
\name{aviodebris-package}
\alias{aviodebris}
\alias{aviodebris-package}
\title{aviodebris: Estimate impact of space debris reentries on aviation}
\description{
Functions to help prepare the data and plots for the DASC 2025 paper "Preparing for Potential Closure of European Airspaces due to Re-entering Space Objects".
}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
