% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_spatial.R
\name{bbox_of_hexes_for_bbox_at_res}
\alias{bbox_of_hexes_for_bbox_at_res}
\title{Return the bounding box of the H3 hexagons covering a user bounding box.}
\usage{
bbox_of_hexes_for_bbox_at_res(bbox, resolution)
}
\arguments{
\item{bbox}{the user bounding box}

\item{resolution}{the resolution of the H3 hexagons covering `bbox`}
}
\value{
the bounding box containing the H3 hexagons covering `bbox`
}
\description{
Return the bounding box of the H3 hexagons covering a user bounding box.
}
