% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effective_exposed_area}
\alias{effective_exposed_area}
\title{Effective Exposed Area}
\format{
A data frame with columns:
\itemize{
\item \code{icao}: the ICAO aircraft type code.
\item \code{eea}: the effective exposed area.
\item \code{pax}: the max number of passengers.
}
}
\usage{
effective_exposed_area
}
\description{
Effective Exposed Area
}
\keyword{datasets}
