% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_traffic}
\alias{resample_traffic}
\title{reasample a day of trajectories at 30s}
\usage{
resample_traffic(day)
}
\arguments{
\item{day}{the date for the trajectories, it refers to the relevant
parquet file in `data-raw/trjs/trjs_<YYYY-MM-DD>.parquet`}
}
\value{
a data frame of resampled points is saved in
         `data/trajectories_<YYYY-MM-DD>_resampled_30s.parquet`
}
\description{
reasample a day of trajectories at 30s
}
