% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aircraft_performance_scrapers.R
\name{scrape_aircraft_type_info_doc8643}
\alias{scrape_aircraft_type_info_doc8643}
\title{Scraper for `doc8643.com`'s Aircraft info}
\usage{
scrape_aircraft_type_info_doc8643(ac_type, session)
}
\arguments{
\item{ac_type}{ICAO aircraft type}

\item{session}{`polite` session}
}
\value{
a dataframe of aircraft info
}
\description{
Scraper for `doc8643.com`'s Aircraft info
}
\examples{
\dontrun{
host <- "https://doc8643.com/"
session <- polite::bow(host, force = TRUE)
scraper <- purrr::partial(scrape_aircraft_type_info_doc8643, session = session)

ac_types |>
  map(.f = scraper) |>
  bind_rows() |>
  # fix some data
  mutate(
    type = case_when(
      icao == "C700" ~ "L2J",
      icao == "C68A" ~ "L2J",
      .default = .data$type
    )
  ) |>
  separate_wider_delim(
    manufacturer,
    " ",
    names = c("manufacturer", "other"),
    too_many = "drop"
  ) |>
  # fix some data
  mutate(
    manufacturer = case_when(
      manufacturer == "AIR" ~ str_c(manufacturer, other, sep = " "),
      manufacturer == "AVIONES" ~ str_c(manufacturer, other, sep = " "),
      manufacturer == "GULFSTREAM" ~ str_c(manufacturer, other, sep = " "),
      other == "HELICOPTERS" & manufacturer == "AIRBUS" ~
        str_c(manufacturer, other, sep = " "),
      other == "GRUMMAN" & manufacturer == "NORTHROP" ~
        str_c(manufacturer, other, sep = " "),
      other == "MARTIN" & manufacturer == "LOCKHEED" ~
        str_c(manufacturer, other, sep = " "),
      .default = manufacturer
    )
  ) |>
  select(-other)
}
}
