% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aircraft_performance_scrapers.R
\name{scrape_aircraft_type_info_perfdb}
\alias{scrape_aircraft_type_info_perfdb}
\title{Scraper for EUROCONTROL's Aircraft Performance DB}
\usage{
scrape_aircraft_type_info_perfdb(ac_type, session)
}
\arguments{
\item{ac_type}{ICAO aircraft type}

\item{session}{`polite` session}
}
\value{
a dataframe of aircraft info
}
\description{
Scraper for EUROCONTROL's Aircraft Performance DB
}
\examples{
\dontrun{
host <- "https://contentzone.eurocontrol.int/"
session <- polite::bow(host, force = TRUE)
scraper <- purrr::partial(scrape_aircraft_type_info_perfdb,
                          session = session)

ac_types |>
  map(.f = scraper) |>
  bind_rows()
}
}
