% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_density.R
\name{traffic_density_hourly}
\alias{traffic_density_hourly}
\title{Calculated the hourly traffic density per aircraft type}
\usage{
traffic_density_hourly(days, resolution = 3L, interval = 30L)
}
\arguments{
\item{days}{a vector of dates which correspond to parquet files
in `data/trajectories_<YYYY-MM-DD>_resampled_<interval>s_bbox_res_<resolution>.parquet`}

\item{resolution}{the resolution of the H3 hexagonal cells}

\item{interval}{resampling interval}
}
\value{
save the result in a parquet file `traffic_density_hourly.parquet`
}
\description{
Calculated the hourly traffic density per aircraft type
}
