% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weightings_h3_resolution_3_hourly}
\alias{weightings_h3_resolution_3_hourly}
\title{Debris hourly density function per H3 hex cell at resolution 3}
\format{
## `weightings_half_degree_hourly`
A data frame with 360 * 720 rows and 1 column:
* `h3_resolution_3`: the H3 cell id
* `w`: the value of the hourly weighting function.
}
\usage{
weightings_h3_resolution_3_hourly
}
\description{
Hourly weighting functions for reentered debris in 25 years from
2000-01-01 till 2025-01-01 (excluded), for H3 cells at resolution 3.
}
\keyword{datasets}
