% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weightings_half_degree_hourly}
\alias{weightings_half_degree_hourly}
\title{Debris weighting function for half degree grid}
\format{
## `weightings_half_degree_hourly`
A data frame with 360 * 720 rows and 3 columns:
* `lon`: the longitude.
* `lat`: the latitude.
* `val`: the value of the hourly weighting function.
}
\usage{
weightings_half_degree_hourly
}
\description{
Hourly weighting functions for reentered debris in 25 years from
2000-01-01 till 2025-01-01 (excluded), at half degree resolution
worldwide.
}
\keyword{datasets}
