% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weightings_half_degree_latitude}
\alias{weightings_half_degree_latitude}
\title{Debris weighting function per latitude}
\format{
A data frame with 360 rows (latitude from -90 to 90 at half degree)
and 2 columns:
* `lat`: the latitude.
* `val`: the value of the weighting function.
}
\usage{
weightings_half_degree_latitude
}
\description{
weighting functions for reentered debris in 25 years from
2000-01-01 till 2025-01-01 (excluded), at half degree resolution.
}
\keyword{datasets}
